+++ src/FormControl/test/FormControlSpec.tsx
@@ -533,190 +533,6 @@ describe('FormControl', () => {
 
       expect(formValue.user.name[1]).to.equal('tom');
     });
-
-    it('Should render an error message when the value changes', () => {
-      const model = Schema.Model({
-        user: Schema.Types.ObjectType().shape({
-          age: Schema.Types.NumberType('Age must be a number ')
-        })
-      });
-
-      render(
-        <Form formDefaultValue={{ user: { age: '10' } }} model={model} nestedField>
-          <FormControl name="user.age" />
-        </Form>
-      );
-
-      fireEvent.change(screen.getByRole('textbox'), { target: { value: 'a' } });
-
-      expect(screen.getByRole('alert')).to.have.text('Age must be a number ');
-    });
-
-    it('Should render an error message when the value changes with checkAsync', async () => {
-      const model = Schema.Model({
-        user: Schema.Types.ObjectType().shape({
-          age: Schema.Types.NumberType('Age must be a number ')
-        })
-      });
-
-      render(
-        <Form formDefaultValue={{ user: { age: '10' } }} model={model} nestedField>
-          <FormControl checkAsync name="user.age" />
-        </Form>
-      );
-
-      fireEvent.change(screen.getByRole('textbox'), { target: { value: 'a' } });
-      await screen.findByRole('alert');
-      expect(screen.getByRole('alert')).to.have.text('Age must be a number ');
-    });
-
-    it('Should render an error message when the value changes', () => {
-      const model = Schema.Model({
-        age: Schema.Types.NumberType('Age must be a number ')
-      });
-
-      render(
-        <Form formDefaultValue={{ age: '10' }} model={model} nestedField>
-          <FormControl name="age" />
-        </Form>
-      );
-
-      fireEvent.change(screen.getByRole('textbox'), { target: { value: 'a' } });
-
-      expect(screen.getByRole('alert')).to.have.text('Age must be a number ');
-    });
-
-    it('Should render an error message when the form is checked', async () => {
-      const model = Schema.Model({
-        user: Schema.Types.ObjectType().shape({
-          age: Schema.Types.NumberType('Age must be a number ')
-        })
-      });
-
-      const ref = React.createRef<any>();
-
-      render(
-        <Form
-          formDefaultValue={{ user: { age: '10' } }}
-          model={model}
-          nestedField
-          ref={ref}
-          checkTrigger="none"
-        >
-          <FormControl name="user.age" />
-        </Form>
-      );
-
-      fireEvent.change(screen.getByRole('textbox'), { target: { value: 'a' } });
-
-      ref.current.check();
-
-      await waitFor(() => {
-        expect(screen.getByRole('alert')).to.have.text('Age must be a number ');
-      });
-    });
-
-    it('Should only validate the field of the current interactive control', () => {
-      const onCheck = sinon.spy();
-      const model = Schema.Model({
-        obj: Schema.Types.ObjectType().shape({
-          number1: Schema.Types.NumberType(),
-          number2: Schema.Types.NumberType()
-        })
-      });
-
-      render(
-        <Form
-          formDefaultValue={{ obj: { number1: '', number2: '' } }}
-          model={model}
-          nestedField
-          onCheck={onCheck}
-        >
-          <FormControl name="obj.number1" data-testid="textbox1" />
-          <FormControl name="obj.number2" data-testid="textbox2" />
-        </Form>
-      );
-
-      fireEvent.change(screen.getByTestId('textbox1'), { target: { value: 'a' } });
-
-      expect(screen.getByText('obj.number1 must be a number')).to.exist;
-
-      fireEvent.change(screen.getByTestId('textbox2'), { target: { value: 'a' } });
-
-      expect(screen.getByText('obj.number2 must be a number')).to.exist;
-
-      fireEvent.change(screen.getByTestId('textbox2'), { target: { value: 1 } });
-
-      expect(screen.queryByText('obj.number2 must be a number')).to.not.exist;
-      expect(screen.queryByText('obj.number1 must be a number')).to.exist;
-    });
-
-    it('Should render deeply nested error messages', async () => {
-      const model = Schema.Model({
-        name: Schema.Types.StringType().isRequired('Name is required'),
-        user: Schema.Types.ObjectType().shape({
-          email: Schema.Types.StringType().isRequired('Email is required'),
-          city: Schema.Types.StringType().isRequired('City is required')
-        })
-      });
-
-      render(
-        <Form model={model} nestedField>
-          <FormControl name="name" />
-          <FormControl name="user.email" />
-          <FormControl name="user.city" />
-          <Button appearance="primary" type="submit">
-            Submit
-          </Button>
-        </Form>
-      );
-
-      fireEvent.click(screen.getByRole('button'));
-
-      await screen.findAllByRole('alert');
-
-      expect(screen.getAllByRole('alert')).to.have.length(3);
-
-      screen.getAllByRole('alert').forEach((alert, index) => {
-        expect(alert).to.have.text(
-          ['Name is required', 'Email is required', 'City is required'][index]
-        );
-      });
-    });
-
-    it('Should render deeply nested error messages with FormControl set rule', async () => {
-      render(
-        <Form nestedField>
-          <FormControl
-            name="name"
-            rule={Schema.Types.StringType().isRequired('Name is required')}
-          />
-          <FormControl
-            name="user.email"
-            rule={Schema.Types.StringType().isRequired('Email is required')}
-          />
-          <FormControl
-            name="user.city"
-            rule={Schema.Types.StringType().isRequired('City is required')}
-          />
-          <Button appearance="primary" type="submit">
-            Submit
-          </Button>
-        </Form>
-      );
-
-      fireEvent.click(screen.getByRole('button'));
-
-      await screen.findAllByRole('alert');
-
-      expect(screen.getAllByRole('alert')).to.have.length(3);
-
-      screen.getAllByRole('alert').forEach((alert, index) => {
-        expect(alert).to.have.text(
-          ['Name is required', 'Email is required', 'City is required'][index]
-        );
-      });
-    });
   });
 
   describe('Accessibility', () => {
